add timer and intro

+++ src/vga/vga.h
@@ -30,7 +30,7 @@ void put_pixel(PointScreen_t point, u8 color);
 void draw_line(PointScreen_t a, PointScreen_t b, u8 color);
 void draw_rectangular(PointScreen_t p1, PointScreen_t p2, u8 color);
 void draw_character(char c, PointScreen_t start_point, int color);
-void draw_string(char *str, PointScreen_t start_point, int color);
+void draw_string(char *str, PointScreen_t start_point, uint8_t color);
 void render_screen();
 
 #endif
